Completely new structure #22
Much less manual work, in case we want to re-run all graphs, updated with calibration of mass flow controllers, final values for water content and porosity and pump calibration. And the new structure is (possibly?) easier to understand, due to less manual work.
Still missing compared to the original structure: Plotting of profiles and allowing to tweak parameters. Experiment 1, 7 and 8 is still in the original structure, due to having a different setup than the rest.