Reaction in the reservoir and possibility to change kl,kg and ae separately
We can now set Kga ti 'individual' and add kg,kl and ae values. Also made a loop for changing on parameter at a time, and tried changing k and Kga for the pH 6 graph, and ae for the pH 6 and 7 graphs. Added all profiles for standard experiments (2-6). Put automatic optimization in a separate task.